requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/dev.in
#
asttokens==2.2.0
    # via stack-data
backcall==0.2.0
    # via ipython
build==0.9.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via pip-tools
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
distlib==0.3.6
    # via virtualenv
executing==1.2.0
    # via stack-data
filelock==3.8.0
    # via virtualenv
identify==2.5.9
    # via pre-commit
ipdb==0.13.9
    # via -r requirements/dev.in
ipython==8.7.0
    # via ipdb
jedi==0.18.2
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
nodeenv==1.7.0
    # via pre-commit
packaging==21.3
    # via build
parso==0.8.3
    # via jedi
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.11.0
    # via -r requirements/dev.in
platformdirs==2.5.4
    # via virtualenv
pre-commit==2.20.0
    # via -r requirements/dev.in
prompt-toolkit==3.0.33
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.13.0
    # via ipython
pyparsing==3.0.9
    # via packaging
python-decouple==3.6
    # via -r requirements/dev.in
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via asttokens
stack-data==0.6.2
    # via ipython
toml==0.10.2
    # via
    #   ipdb
    #   pre-commit
tomli==2.0.1
    # via
    #   build
    #   pep517
traitlets==5.6.0
    # via
    #   ipython
    #   matplotlib-inline
virtualenv==20.17.0
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
